# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Exam Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List questions
      parameters:
        - name: examId
          in: query
          schema:
            type: integer
        - name: sectionId
          in: query
          schema:
            type: integer
        - name: subjectId
          in: query
          schema:
            type: integer
        - name: topicId
          in: query
          schema:
            type: integer
        - name: difficulty
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Created question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get question by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Update question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionUpdate"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # TOPIC MOCKS
...
  /quizzes:
    get:
      operationId: listQuizzes
      tags: [quizzes]
      summary: List quizzes
      parameters:
        - name: quizType
          in: query
          schema:
            type: string
        - name: published
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
    post:
      operationId: createQuiz
      tags: [quizzes]
      summary: Create quiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizInput"
      responses:
        "201":
          description: Created quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{id}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      summary: Get quiz by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
    patch:
      operationId: updateQuiz
      tags: [quizzes]
      summary: Update quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizUpdate"
      responses:
        "200":
          description: Updated quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
    delete:
      operationId: deleteQuiz
      tags: [quizzes]
      summary: Delete quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /quizzes/{id}/publish:
    patch:
      operationId: toggleQuizPublish
      tags: [quizzes]
      summary: Toggle quiz publish status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"

  # ATTEMPTS
...
components:
  schemas:
...
    Question:
      type: object
      required: [id, questionText, optionA, optionB, optionC, optionD, correctAnswer, marks, createdAt]
      properties:
        id:
          type: integer
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: string
        explanation:
          type: ["string", "null"]
        marks:
          type: number
        negativeMarks:
          type: number
        difficulty:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        examId:
          type: ["integer", "null"]
        sectionId:
          type: ["integer", "null"]
        subjectId:
          type: ["integer", "null"]
        topicId:
          type: ["integer", "null"]
        quizId:
          type: ["integer", "null"]
        topicMockId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    QuestionInput:
      type: object
      required: [questionText, optionA, optionB, optionC, optionD, correctAnswer, marks]
      properties:
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: string
        explanation:
          type: string
        marks:
          type: number
        negativeMarks:
          type: number
        difficulty:
          type: string
        year:
          type: integer
        examId:
          type: integer
        sectionId:
          type: integer
        subjectId:
          type: integer
        topicId:
          type: integer
        quizId:
          type: integer
        topicMockId:
          type: integer
    QuestionUpdate:
      type: object
      properties:
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: string
        explanation:
          type: string
        marks:
          type: number
        negativeMarks:
          type: number
        difficulty:
          type: string
        year:
          type: integer
        examId:
          type: integer
        sectionId:
          type: integer
        subjectId:
          type: integer
        topicId:
          type: integer
        quizId:
          type: integer
        topicMockId:
          type: integer

    # TOPIC MOCKS
...
    Quiz:
      type: object
      required: [id, title, quizType, duration, totalMarks, published, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        quizType:
          type: string
        duration:
          type: integer
        totalMarks:
          type: integer
        negativeMarks:
          type: number
        published:
          type: boolean
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        questionCount:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    QuizInput:
      type: object
      required: [title, quizType, duration, totalMarks]
      properties:
        title:
          type: string
        quizType:
          type: string
        duration:
          type: integer
        totalMarks:
          type: integer
        negativeMarks:
          type: number
        published:
          type: boolean
        scheduledAt:
          type: string
          format: date-time
    QuizUpdate:
      type: object
      properties:
        title:
          type: string
        quizType:
          type: string
        duration:
          type: integer
        totalMarks:
          type: integer
        negativeMarks:
          type: number
        published:
          type: boolean
        scheduledAt:
          type: string
          format: date-time

    # ATTEMPTS
...